requirements.txt
streamlit
requests
python-dotenv
azure-cognitiveservices-speech
pydub
